updated the Readme for Kotlin
