+++ kotlin/services/secrets-manager/build.gradle.kts
